Refactor: Decouple MyMessage specifics from cell
